Setup load tests for new K8s staging env
We are in the process of migrating Prison Visits from template deploy to
Kubernetes.  Part of this piece of work is ensuring that all current
tests are run against the new environment so we can be confident with
any future deployments.  This PR creates a new load test repo that
points to the new staging URL and will, for the time being, be run
alongside the existing load tests.  Once the migration has been
completed we can clean up/remove the old tests on Jenkins and related
repos on Github.